third_party/WebKit/LayoutTests/http/tests/security/cannot-read-cssrules-expected.txt

This test whether a script can read the rules from a cross-origin style
sheet.  For more information on why we block this, please seehttps://bugs.webkit.org/show_bug.cgi?id=20527.

Test begins.
== Cross-Origin ==
cssRules: null
rules: null
== Same-Origin ==
cssRules: [object CSSRuleList]
rules: [object CSSRuleList]
Test ends.
